tests: move to authnetication test class
